SyncOffsetFile moved to core components
