+++ tests/unit/neptune/new/cli/utils.py
@@ -23,11 +23,11 @@ from neptune.constants import (
     ASYNC_DIRECTORY,
     OFFLINE_DIRECTORY,
 )
+from neptune.core.components.queue.sync_offset_file import SyncOffsetFile
 from neptune.exceptions import MetadataContainerNotFound
 from neptune.internal.backends.api_model import ApiExperiment
 from neptune.internal.container_type import ContainerType
 from neptune.internal.disk_queue import DiskQueue
-from neptune.internal.utils.sync_offset_file import SyncOffsetFile
 from tests.unit.neptune.new.utils.api_experiments_factory import (
     api_metadata_container,
     api_run,
